refactor: 사용하지 않는 코드 삭제 및 정리(주석 처리)
