+++ src/app/(public)/mate/_components/post/postForm.tsx
@@ -1,12 +1,11 @@
 "use client";
 
 import { useState } from "react";
-import { useQueryClient, useMutation, useQuery } from "@tanstack/react-query";
+import { useQueryClient, useMutation } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import dynamic from "next/dynamic";
 import { locationStore } from "@/zustand/locationStore";
 import { useAuthStore } from "@/zustand/useAuth";
-import { getConvertAddress } from "../../getConvertAddress";
 import { getConvertDate } from "../../../../utils/getConvertDate";
 import Swal from "sweetalert2";
 import PetForm from "./pet/petForm";
@@ -38,19 +37,10 @@ const PostForm = () => {
     pet_id: []
   };
 
-  // const initialPetState: Pets = {
-  //   userId,
-  //   pet_id: []
-  // };
-
   const [formPosts, setFormPosts] = useState<Omit<MateNextPostType, "user_id">>(initialState);
-  // const [formPets, setFormPets] = useState<Pets[]>([initialPetState]);
-
-  // console.log(formPets);
 
   // 게시물 등록
   const addPost = async (formAllData: { post: MateNextPostType }) => {
-    // console.log("데이터 넘어오는 거 확인", formAllData);
     try {
       const response = await fetch(`/api/mate`, {
         method: "POST",
@@ -59,18 +49,14 @@ const PostForm = () => {
         },
         body: JSON.stringify({
           post_data: formAllData.post
-          // pets_data: formAllData.pets
         })
       });
-
-      // console.log("Response status:", response.status); // 응답 상태 로그
 
       if (!response.ok) {
         throw new Error(`HTTP error! status: ${response.status}`);
       }
 
       const data = await response.json();
-      // console.log("Response data:", data);
 
       return data;
     } catch (error) {
@@ -103,7 +89,6 @@ const PostForm = () => {
     e.preventDefault();
 
     if (!isFormValid()) {
-      // alert("모든 항목을 입력해 주세요!");
       Swal.fire({
         title: "모든 항목을 입력해 주세요!",
         text: "빠진 부분이 있는지 확인해 주세요.",
@@ -128,8 +113,6 @@ const PostForm = () => {
       console.error(err);
     }
   };
-
-  // console.log(userPets)
 
   return (
     <div className="min-h-screen">
@@ -233,7 +216,7 @@ const PostForm = () => {
           <p className="flex justify-end text-subTitle2">{formPosts.content?.length}/200</p>
         </div>
         {/* 반려동물 정보 등록 */}
-        <PetForm setFormPosts={setFormPosts} userId={userId} />
+        <PetForm setFormPosts={setFormPosts} />
         {/* 작성하기 버튼 */}
         <div className="mb-[7.5rem] mt-[1.5rem] flex w-full items-center justify-center px-[1.5rem]">
           <button
